Add error log extraction for improved debugging output
