+++ helpers/logger.ts
@@ -161,10 +161,54 @@ export interface TestLogOptions {
   enableLogging: boolean;
   customLogFile?: string;
   testName: string;
+  logFileName?: string;
 }
 
+// Extract error logs from log files
+export function extractErrorLogs(testName: string): string {
+  if (!fs.existsSync("logs")) {
+    return "";
+  }
+  console.log("testName", testName);
+
+  try {
+    const logFiles = fs
+      .readdirSync("logs")
+      .filter((file) => file.endsWith(".log") && file.includes(testName));
+    const errorLines: string[] = [];
+
+    for (const logFile of logFiles) {
+      const logPath = path.join("logs", logFile);
+      const content = fs.readFileSync(logPath, "utf-8");
+      const lines = content.split("\n");
+
+      for (const line of lines) {
+        if (/fail/i.test(line)) {
+          let lineToAdd = line;
+          lineToAdd = lineToAdd?.split(">")[1]?.trim();
+          lineToAdd = lineToAdd?.split("//")[0]?.trim();
+          lineToAdd = lineToAdd?.replace("expected false to be true", "failed");
+          if (lineToAdd) {
+            errorLines.push(lineToAdd);
+          }
+        }
+      }
+    }
+
+    console.log(errorLines);
+    if (errorLines.length > 0) {
+      return `\n\n*Error Logs:*\n\`\`\`\n${errorLines.slice(-5).join("\n")}\n\`\`\``;
+    }
+  } catch (error) {
+    console.error("Error reading log files:", error);
+  }
+
+  return "";
+}
 export const createTestLogger = (options: TestLogOptions) => {
   let logStream: fs.WriteStream | undefined;
+  // Extract clean test name for log file (remove path and extension)
+  let logFileName: string = "";
 
   if (options.enableLogging) {
     // Ensure logs directory exists
@@ -173,8 +217,6 @@ export const createTestLogger = (options: TestLogOptions) => {
       fs.mkdirSync(logsDir, { recursive: true });
     }
 
-    // Extract clean test name for log file (remove path and extension)
-    let logFileName: string;
     if (options.customLogFile) {
       logFileName = options.customLogFile;
     } else {
@@ -212,7 +254,11 @@ export const createTestLogger = (options: TestLogOptions) => {
     }
   };
 
-  return { processOutput, close };
+  return {
+    processOutput,
+    close,
+    logFileName,
+  };
 };
 
 // Optional: Add file logging capability
